🚚 Relocate to common directory
